Workaround class-variables inheritance
As documented in manastech/crystal#916, cannot use class variables
and inheritance properly.

To workaround that issue, introduce `routing` macro which defines
a instance method used to populate the routing table of each
Router instance.

This of course has a drawback: new instances of same router will
consume larger amount of memory due duplicated routes, but is
enough for now as workaround.