Refactor parameter extraction and pattern matching
Use dedicated `CompiledPattern` utility object to deal with these
two.

This reduces the exposed API of `Route`.

Also introduce `Action.call` to allow `Action` define any kind
of initializer that is not tied to a specific request.

This allow instantiate an action and test it totally isolated
of the request.

If necessary due any helper, `request` property is accessible.