benchmarks: compare performance of some web frameworks
Setup a simple test to compare a few Crystal-powered web frameworks:
Amethyst, Artemis and Beryl (this project)

The idea of this comparison is to serve as foundation for performance
and optimization analysis.

Run benchmarks of each framework against several scenarios:

- Root (/)
- Named segment (/posts/:id, to test processing of params)
- Middle request (pick the middle endpoint of a long list)
- Bottom request (pick the last route of a long list)

Results obtained from a simple run on OSX:

    amethyst (req_root) 17874.15 (± 150.41) 24.81× slower
     artanis (req_root) 443421.31 (± 5738.80)       fastest
       beryl (req_root) 128771.06 (± 9765.26)  3.44× slower
    amethyst (req_id) 15667.19 (± 364.56) 16.22× slower
     artanis (req_id) 254103.97 (± 3104.89)       fastest
       beryl (req_id) 69536.91 (± 1022.40)  3.65× slower
    amethyst (req_middle) 13556.67 (± 107.90)  4.52× slower
     artanis (req_middle) 61224.30 (± 411.86)       fastest
       beryl (req_middle)  4180.92 (± 32.31) 14.64× slower
    amethyst (req_bottom)  9925.64 (± 88.15)  6.90× slower
     artanis (req_bottom) 68454.69 (± 659.62)       fastest
       beryl (req_bottom)  2847.93 (± 19.47) 24.04× slower

Shows Beryl underperforms on every scenario. This could be
caused by performing Regexp comparison and also usage of
Struct vs Classes when passing Routes and CompiledPattern
between calls.

Last but not least, it could be related to `params` extraction
from request, which might indicate perhaps lazy evaluation of
those might be recommended.

[skip ci]